+++ server/modules/balance.js
@@ -2,45 +2,41 @@ const {fetchGraphQL, queryBalance} = require("./graphql");
 const {parseTime} = require("../helpers/parsers");
 const {addressBalance} = require("../helpers/node-data");
 const {SYNC_STATE_SYNCED, SYNC_STATE_CATCHUP} = require("../helpers/consts");
-const {sendAlert, sendMessage} = require("../helpers/messangers");
+const {sendMessage} = require("../helpers/messangers");
+const {isset} = require("../helpers/isset");
+const {shortAddress} = require("../helpers/short-address")
 
 const processBalance = async () => {
-    const {publicKey, nodeInfoCollectInterval = "30s", graphql} = globalThis.config
+    const {publicKey = "", publicKeyDelegators = "", nodeInfoCollectInterval = "30s", graphql} = globalThis.config
     const _nodeInfoCollectInterval = parseTime(nodeInfoCollectInterval)
+    const address = publicKey ? publicKey : publicKeyDelegators ? publicKeyDelegators : null
+
+    if (!address) return
 
     try {
-        globalThis.cache.balance = publicKey ? await fetchGraphQL(graphql, queryBalance, {publicKey}) : 0
+        const balanceData = await fetchGraphQL(graphql, queryBalance, {publicKey: address})
+
+        if (balanceData) {
+            const {total = 0, liquid = 0, locked = 0} = addressBalance(balanceData)
+            const shortAddr = shortAddress(address, 5)
+            const totalValue = (total/10**9).toFixed(4), liquidValue = (liquid/10**9).toFixed(4), lockedValue = (locked/10**9).toFixed(4)
+
+            if (cache.balance) {
+                const {total: cacheTotal = 0, liquid: cacheLiquid = 0, locked: cacheLocked = 0} = addressBalance(cache.balance)
+                if (+(cacheTotal) !== +(total)) {
+                    sendMessage("BALANCE", `Balance for address \`${shortAddr}\` was changed to \`${(total / 10**9).toFixed(4)}\` (Total), \`${(liquid / 10**9).toFixed(4)}\` (Movable), \`${(locked / 10**9).toFixed(4)}\` (Locked).`)
+                }
+            } else {
+                sendMessage("BALANCE", `Balance state for address \`${shortAddr}\` is \`${(total / 10**9).toFixed(4)}\` (Total), \`${(liquid / 10**9).toFixed(4)}\` (Movable), \`${(locked / 10**9).toFixed(4)}\` (Locked).`)
+            }
+
+            globalThis.cache.balance = balanceData
+        }
     } catch (e) {}
 
     setTimeout(processBalance, _nodeInfoCollectInterval)
 }
 
-const processBalanceSend = async () => {
-    if (!globalThis.config || !globalThis.config.publicKey) return
-
-    const {balanceSendInterval} = globalThis.config
-    const _balanceSendInterval = parseTime(balanceSendInterval)
-
-    let reload
-
-    let balance = addressBalance(globalThis.cache.balance)
-    let daemon = globalThis.cache.daemon
-
-    if (balance && daemon && [SYNC_STATE_SYNCED, SYNC_STATE_CATCHUP].includes(daemon.syncStatus)) {
-        const {total, liquid, locked} = balance
-        const message =`Balance state: \`${(total / 10**9).toFixed(4)}\` (Total), \`${(liquid / 10**9).toFixed(4)}\` (Movable), \`${(locked / 10**9).toFixed(4)}\` (Locked).`
-
-        globalThis.currentBalance = total
-        sendMessage("BALANCE", message)
-        reload = _balanceSendInterval
-    } else {
-        reload = 5000
-    }
-
-    setTimeout(processBalanceSend, reload)
-}
-
 module.exports = {
-    processBalance,
-    processBalanceSend
+    processBalance
 }