+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 + [x] Server: to config parameters `alertToTelegram`, `alertToDiscrod` added value `PRICE` 
 + [x] Server: to config parameters `alertToTelegram`, `alertToDiscrod` added value `REWARDS`
 + [x] Server: to config parameters `alertToTelegram`, `alertToDiscrod` added value `UPTIME`
++ [x] Server: removed config parameter `balanceSendInterval`
 
 ### 2.0.3
 + [x] Server: added support for time interval string values for parameter `restartAfterNotSynced`